app/konradgnat/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
bleach==3.3.0
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
Django==3.1.4
django-cors-headers==3.6.0
django-markdownify==0.9.0
django-taggit==1.3.0
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
Markdown==3.3.4
packaging==20.9
pyparsing==2.4.7
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
virtualenv==20.2.2
webencodings==0.5.1
waitress==2.1.2
whitenoise==6.5.0
